Revert exception-moving
