+++ mlflow/utils/exception.py
@@ -1,0 +1,7 @@
+class MLflowException(Exception):
+    pass
+
+
+class ExecutionException(Exception):
+    """Exception thrown when executing a project fails."""
+    pass
