+++ tests/projects/test_projects.py
@@ -9,7 +9,7 @@ import pytest
 
 import mlflow
 from mlflow.entities.run_status import RunStatus
-from mlflow.exceptions import ExecutionException
+from mlflow.utils.exception import ExecutionException
 from mlflow.store.file_store import FileStore
 from mlflow.utils import env
